// src/types/mock-data.type.ts
import { MockCity } from './mock-city.type.js';

export type MockData = {
  titles: string[];
  descriptions: string[];
  cities: MockCity[];
  previewImage: string;
  fullImage: string;
  types: string[];
  goods: string[];
  names: string[];
  emails: string[];
  avatar: string;
  userTypes: string[];
}